+++ js/photos.js
@@ -1,0 +1,28 @@
+const photoTemplate = document.querySelector('#picture')
+  .content
+  .querySelector('.picture');
+
+const container = document.querySelector('.pictures');
+
+const createFoto = ({ comments, description, likes, url}) => {
+  const photo = photoTemplate.cloneNode(true);
+
+  photo.querySelector('.picture__img').src = url;
+  photo.querySelector('.picture__img').alt = description;
+  photo.querySelector('.picture__likes').textContent = likes;
+  photo.querySelector('.picture__comments').textContent = comments.length;
+
+  return photo;
+};
+
+export const renderPhoto = (pictures) => {
+  const fragment = document.createDocumentFragment();
+
+  pictures.forEach((picture) => {
+    const photo = createFoto(picture);
+
+    fragment.append(photo);
+  });
+
+  container.append(fragment);
+};
